+++ backend/StudentManagement.Application/DTOs/Instructor/InstructorDto.cs
@@ -2,7 +2,7 @@ namespace StudentManagement.Application.DTOs.Instructor;
 
 public class InstructorDto
 {
-    public int InstructorSsn { get; set; }
+    public long InstructorSsn { get; set; }
     public string FirstName { get; set; } = string.Empty;
     public string LastName { get; set; } = string.Empty;
     public string? Phone { get; set; }
